refactor(download_history): Enhance thread safety and code organization
- Extract JSON serialization logic into a separate file (json_serializer.go)
- Move type definitions to types.go for better code organization
- Add thread-safe counter implementation in counter.go
- Improve error handling and state management
- Add GetStats() method that doesn't return an error
- Update ExportStats to be a simple struct without methods
- Fix race conditions in concurrent access patterns
